requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.10.0
Flask-WTF==1.2.2
python-dotenv==1.0.1
